Figuring out how to render dropdown menu on title submission
