+++ client/src/components/Main/Song.jsx
@@ -4,8 +4,10 @@ import React from 'react';
 const Song = (props) => (
 
       <div>
+          
           {console.log('Song props:',props)}
-          {props.title} by {props.artist}
+            {props.title} by {props.artist}
+          <button onClick={props.dropdownSongs}></button>
       </div>
