+++ client/src/components/Main/SongList.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Song from './Song.jsx'
+
+const SongList = (props) => (
+      <div>
+    {/* {console.log('HERE ARE MY SONG LIST PROPS: ', props)} */}
+        {props.songBank.map( (song,index)=> <Song title={song.title} artist={song.artist} key={index} dropdownSongs={props.dropdownSongs}  / >)}
+      </div>
+  
+    
+)
+
+
+
+export default SongList
